Add files via upload
Search archived security logs on multiple domain controllers for eventIDs and userIDs of interest. 

The search query is built in an here-array that is then passed to the get-winevent command as a -filterXML switch. 

This log can process 5.7 GB of logfiles across three domain controllers in under three minutes. Anybody who has ever tried using the Windows Event Viewer to find anything in the security log of an active domain controller will know how that compares.

The query and the data returned can be built from the XML view of a log event, which can be viewed in the Windows Event Viewer, under the "Details" tab.